Python3/Tornado/apps/ExchangeWalletApi/ExWallet/requirements.txt
asn1crypto==0.24.0
cashaddress==1.0.6
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
coincurve==12.0.0
cytoolz==0.10.0
eth-hash==0.2.0
eth-typing==2.1.0
eth-utils==1.7.0
future==0.17.1
idna==2.8
mypy-extensions==0.4.1
pbkdf2==1.3
py-ecc==1.7.1
pycparser==2.19
PyMySQL==0.9.3
requests==2.22.0
toolz==0.10.0
tornado==6.0.3
urllib3==1.25.6
libeospy==2.0.1
stellar-sdk==2.1.0
monero==0.7.1